feat: split into view-hooks
